Wrote basic i2c functions to access the battery babysitter data we are interested in.
Other than looking at the device type(which still kinda confuses me), all the code does is
read a specific register in I2C that stores a value we want. I made functions for voltage
current state of charge and temperature with the register to read hard coded into them.
For testing the values, voltage I compared to a multimeter reading the voltage so I
know that one is correct. Current was at 0mA becuase nothing is plugged into it atm.
Capacity seems higher than what it should be so im not sure about that one(might not be
in hex). state of charge was at 50 which seems possible but if I recall it was at
30ish when we were testing with the arduino so I will want to fully charge it then
see if it is at 100% in hex or in int form. Temp was at around 295.2 Kelvin which is around
room temp so that one is correct.