Wrote code to capture runtime power usage per quantum
Wrote what will be a submodule of suneed that will read from the i2c bus
at a constant rate of 5 seconds and update its values for voltage,
current, capacity, and state of charge.

I also wrote my idea for quantum based power measurement where
suneed reads the average current and differece in capacity for a
period of time and updates a shared data structure containing
this information(for now it writes it to a file).