Implement client app identification
When a client sends an IPC request, if it is a new pipe ID (assigned
via nng) a new entry is registered in the client pipe table. This
table is used to identify the "state" of the client pipe; the client
pipe can send various tokens in order to put itself in a state.
When a client pipe is in a state, the next message received from it
will be treated as arguments for an associated command. This is
necessary for sending variable-length data, such as strings, over
IPC.

Next up is finishing the device registration/function linkage
process so that client programs can access devices.