Remove superstition test
The test in the `overlay` suite was basically just me confirming basic behavior of Linux so there's no point in keeping it around. I removed the code associated with it in `sunneed` as well. There's still the build step that builds it with `TESTING` enabled, but that might come in handly later so it can stay.